requirements.txt
python-qpid-proton==0.39.0
marshmallow==3.20.1
url64==0.1.0
docker==6.1.3
boto3==1.28.73
